Restructure all modules
