+++ src/faucet/views/alert.rs
@@ -1,0 +1,77 @@
+use crate::faucet::FaucetController;
+use leptos::prelude::*;
+use leptos::task::spawn_local;
+use leptos::{component, view, IntoView};
+use std::collections::HashSet;
+use std::time::Duration;
+use uuid::Uuid;
+
+const MESSAGE_FADE_AFTER: Duration = Duration::new(3, 0);
+const MESSAGE_REMOVAL_AFTER: Duration = Duration::new(3, 500_000_000);
+
+#[component]
+pub fn ErrorMessages(
+    errors: Vec<(Uuid, String)>,
+    faucet: RwSignal<FaucetController>,
+) -> impl IntoView {
+    let (fading_messages, set_fading_messages) = signal(HashSet::new());
+    view! {
+        <div class="fixed top-4 left-1/2 transform -translate-x-1/2 z-50">
+            {errors
+                .into_iter()
+                .map(|(id, error)| {
+                    spawn_local(async move {
+                        set_timeout(
+                            move || {
+                                set_fading_messages
+                                    .update(|fading| {
+                                        fading.insert(id);
+                                    });
+                            },
+                            MESSAGE_FADE_AFTER,
+                        );
+                        set_timeout(
+                            move || {
+                                set_fading_messages
+                                    .update(|fading| {
+                                        fading.remove(&id);
+                                    });
+                                faucet.get().remove_error_message(id);
+                            },
+                            MESSAGE_REMOVAL_AFTER,
+                        );
+                    });
+
+                    view! {
+                        <div
+                            class=move || {
+                                if fading_messages.get().contains(&id) {
+                                    "opacity-0 transition-opacity bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative mb-2 w-96"
+                                } else {
+                                    "bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative mb-2 w-96"
+                                }
+                            }
+                            role="alert"
+                        >
+                            <span class="block sm:inline">{error}</span>
+                            <span class="absolute top-0 bottom-0 right-0 px-4 py-3">
+                                <svg
+                                    class="fill-current h-6 w-6 text-red-500"
+                                    role="button"
+                                    xmlns="http://www.w3.org/2000/svg"
+                                    viewBox="0 0 20 20"
+                                    on:click=move |_| {
+                                        faucet.get().remove_error_message(id);
+                                    }
+                                >
+                                    <title>Close</title>
+                                    <path d="M14.348 14.849a1.2 1.2 0 0 1-1.697 0L10 11.819l-2.651 3.029a1.2 1.2 0 1 1-1.697-1.697l2.758-3.15-2.759-3.152a1.2 1.2 0 1 1 1.697-1.697L10 8.183l2.651-3.031a1.2 1.2 0 1 1 1.697 1.697l-2.758 3.152 2.758 3.15a1.2 1.2 0 0 1 0 1.698z" />
+                                </svg>
+                            </span>
+                        </div>
+                    }
+                })
+                .collect::<Vec<_>>()}
+        </div>
+    }
+}
