+++ src/utils/error.rs
@@ -1,6 +1,5 @@
-use std::future::Future;
-
 use leptos::prelude::{RwSignal, Update};
+use std::future::Future;
 use uuid::Uuid;
 
 pub async fn catch_all(
